fastapi-sample/pyproject.toml
[project]
name = "fastapi-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dtorannpu", email = "dtorannpu@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "pyjwt[crypto]>=2.11.0",
]

[project.scripts]
fastapi-sample = "fastapi_sample:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["I"] 
